requirements.txt
# Core ML
torch>=2.3.0
transformers>=4.45.0
datasets>=2.20.0
accelerate>=0.33.0
tokenizers>=0.19.0

# Training / RLHF
trl>=0.10.0
peft>=0.12.0

# Scientific computing
numpy>=1.26.0
scipy>=1.13.0
scikit-learn>=1.5.0
pandas>=2.2.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.2
plotly>=5.22.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
rich>=13.7.0
huggingface_hub>=0.24.0
einops>=0.8.0

# Optional: for LLM judge via API
# anthropic>=0.34.0
# openai>=1.40.0
